Move common plugin code to BotPlugin
